+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.3-dev
 
 - Implement component: `modal`, `pagination`, `tabs`.
+- Text color modifiers.
 
 ## 0.3.2